# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: RideMate API
  version: 0.1.0
  summary: Journey sharing and cost sharing for Istanbul.
  description: |
    RideMate connects people whose journeys are already compatible. A driver is
    making the trip regardless; the product finds passengers going the same way
    and lets them share the journey's legitimate costs.

    It is not taxi, ride-hailing, chauffeur or commercial passenger transport,
    and the contract reflects that: there is no fare, no price, no earnings and
    no payout anywhere in it, now or later. A test enforces that vocabulary
    against this file.

    The versioned product API at `/api/v1` carries authentication: a member
    proves possession of a phone number with a one-time passcode and receives a
    short-lived access token plus a rotating refresh token.

    Everything else the product will eventually offer — routes, seat requests,
    trips, reviews, safety — is deliberately absent. This document describes
    what the service serves, not what it intends to.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/auth/otp:
    post:
      operationId: requestPasscode
      tags: [Authentication]
      summary: Request a one-time passcode
      security: []
      description: |
        Sends a passcode to the given number.

        **The response never reveals whether an account exists.** It is `202`
        with the same body in every case, because the alternative turns this
        endpoint into a membership oracle: anyone could learn who uses RideMate
        by submitting numbers and reading the difference. Nothing on the server
        side even looks — an account is resolved only when a passcode is
        verified, so there is no branch here that could differ.

        For the same reason the body carries no challenge identifier, no expiry,
        no resend timestamp and no hint about previous requests. A caller
        learns only that the request was accepted for delivery.

        `202` rather than `200`: delivery is handed to an SMS provider and is
        not complete when this returns.

        Rate limited per number and per address. The two limits are
        indistinguishable in the response, so the count cannot be probed either.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '202':
          description: Accepted for delivery. Says nothing about the number.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpAccepted'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/registrations/otp:
    post:
      operationId: requestRegistrationPasscode
      tags: [Registration]
      summary: Send a registration passcode on a channel
      security: []
      description: |
        Binds the destination to this registration's channel and sends a
        passcode to it.

        **Binding is write-once per channel.** Naming the same destination again
        is a resend and is accepted; naming a DIFFERENT one is `409`
        `channel_already_bound`, and the body says neither destination and never
        whether the bound one was proven. A member who mistyped starts another
        registration — they are cheap, short-lived, and nothing is unique across
        them.

        **The response never reveals whether an account exists**, exactly as
        `POST /api/v1/auth/otp` does not. It is `202` with the same fixed body
        in every case, and nothing on the server side looks: an account is
        consulted only at completion.

        **A registration's passcode is not a sign-in passcode.** It is issued
        into this registration's own scope, so it cannot be spent at
        `POST /api/v1/auth/otp/verify`, a sign-in code cannot prove a
        registration, and another registration that named the same destination
        cannot see it.

        **Rate limits span registrations.** The resend cooldown and the hourly
        cap are counted per destination across every scope, so opening more
        registrations cannot multiply what one address or one handset receives.

        `500` includes a delivery failure: the passcode was recorded but could
        not be handed to a provider, so a retry is subject to the usual
        cooldown. It says nothing about the destination, and a provider adapter
        must keep it that way — surfacing a per-recipient rejection differently
        from a generic failure would make this a deliverability oracle. **Both
        channels fail closed today**, because no provider has been selected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationPasscodeRequest'
      responses:
        '202':
          description: Accepted for delivery. Says nothing about the destination.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpAccepted'
        '401':
          $ref: '#/components/responses/RegistrationCredentialRejected'
        '409':
          $ref: '#/components/responses/RegistrationConflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The access token from `POST /api/v1/auth/otp/verify` or
        `POST /api/v1/auth/refresh`, sent as `Authorization: Bearer <token>`.

        Opaque. It is a lookup key, not a container: clients must not parse it,
        split it, or infer anything from its shape, and none of its structure
        is contract. Short-lived by design — when it expires, exchange the
        refresh token rather than asking the member to sign in again.

        **The refresh token is not a bearer credential** and must never be sent
        in this header. It goes in the body of the refresh request only. See
        that operation for why.
  responses:
...
    ValidationFailed:
      description: The request body is not usable. Field detail is in `details`.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RegistrationCredentialRejected:
      description: |
        The registration credential is missing, malformed, unknown, expired, or
        belongs to a registration that has already produced an account — and, on
        the verify operation, also every way a passcode can fail.

        Which of those it was is deliberately not stated, and the collapse goes
        further than the one on `401` elsewhere. Separating "expired" from
        "already completed" would tell whoever holds a credential whether the
        registration behind it was ever finished, which is completion history
        for a registration they have just shown they cannot advance.

        There is no recovery from this beyond starting again, or — if the
        registration did complete — signing in normally.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RegistrationConflict:
      description: |
        The registration is not in the state the step needed.

        `details.reason` says which, as a stable machine string a client maps to
        its own approved copy. No `details.current_status` accompanies it: a
        registration has no published state, and naming one here would be the
        status endpoint this contract does not have.

        **The vocabulary is narrower than the server's own**, on purpose.
        `account_already_exists` is one string for two internal facts — an
        address already registered and a number already registered — because a
        caller able to tell them apart could aim a registration at an identifier
        pair and read back which half belongs to a member. Collapsing them
        leaves the one thing a client can act on and removes the oracle.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: |
        Too many attempts. Applies per number and per address; the response does
        not say which, so neither count can be probed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: |
        Something failed on the server. For passcode requests this includes a
        delivery failure: the passcode was recorded but could not be handed to
        a provider, so a retry is subject to the usual cooldown.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  headers:
    RequestId:
      description: |
        The correlation id for this request, echoed on every response. If the
        caller supplies a well-formed UUID it is honoured, so a retry can be
        correlated with its original attempt; anything else is replaced.

        This is a transport concern and never a domain identifier.
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: |
                **Developer-facing English. Clients must never display it.**

                The RideMate client owns its own localized copy with Turkish as
                the source language. Server-authored display text would fork
                message ownership across two repositories and two release
                cadences. Map `code` to a client string instead.

                Fixed wording for `not_found` and `method_not_allowed` in
                every environment, and for `internal_error` outside a debug
                build. Framework text for those cases echoes the requested
                path, enumerates the verbs a route accepts, or names an
                internal class and record id — none of which a caller can act
                on, and all of which describe the service to whoever is
                probing it.

                Never contains anything the caller supplied.
            details:
              type: object
              description: |
                Machine-readable detail. Present for validation errors and for
                the refusals that carry a `reason` — seat-request commands and
                trip lifecycle commands; absent everywhere else, so its presence
                is meaningful.

                Field failures are keyed by field name with an array of
                developer-facing messages. Two reserved keys carry a scalar
                instead: `reason`, a stable discriminator for refusals a client
                must handle differently, and `current_status`, the state of a
                request the caller is already entitled to see. No field is ever
                named `reason` or `current_status`, so the two shapes cannot
                collide.
              properties:
                reason:
                  description: |
                    Whichever domain refused. The vocabularies are separate
                    schemas rather than one merged enum, because no domain
                    should gain a say in another's reasons — and `anyOf` rather
                    than `oneOf`, because some strings are deliberately members
                    of more than one: `route_unavailable` in seat requests and
                    trips, `id_already_used` in seat requests and reviews. The
                    same wire string for the same meaning is what keeps a
                    client's mapping simple; `oneOf` would reject exactly those
                    for matching twice.

                    `recurring_route_unsupported` was a third until Phase 16b
                    let a passenger ask for a seat on a named day of a weekday
                    plan. It is a trip reason only now.
                  anyOf:
                    - $ref: '#/components/schemas/SeatRequestRefusalReason'
                    - $ref: '#/components/schemas/TripRefusalReason'
                    - $ref: '#/components/schemas/ReviewRefusalReason'
                    - $ref: '#/components/schemas/RegistrationRefusalReason'
                current_status:
                  $ref: '#/components/schemas/SeatRequestStatus'
              additionalProperties:
                type: array
                items:
                  type: string
            request_id:
              type: string
              format: uuid
              description: Matches the `X-Request-Id` response header.
    ErrorCode:
      type: string
      description: |
        The stable machine contract. Clients map these to their own localized
        copy, so renaming one is a breaking contract change.
      enum:
        - bad_request
        - unauthenticated
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - validation_failed
        - rate_limited
        - internal_error
    OtpRequest:
      type: object
      additionalProperties: false
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 4
          maxLength: 32
          description: |
            The number as the member would write it. `0532 123 45 67`,
            `+90 532 123 45 67` and `905321234567` are the same person and are
            normalized to one canonical E.164 identity server-side.

            A number in full international form is honoured whatever its
            country; the Türkiye default applies only when no country code is
            given.
          examples:
            - '0532 123 45 67'
    OtpAccepted:
      type: object
      additionalProperties: false
      required: [status]
      description: |
        The entire body. It is this small on purpose: anything varying by
        number — a challenge id, an expiry, a resend time — would leak what the
        `202` exists to hide.
      properties:
        status:
          type: string
          enum: [accepted]
...
    RegistrationCredential:
      type: string
      minLength: 1
      maxLength: 512
      description: |
        The credential from `POST /api/v1/registrations`, carried in the BODY of
        every step that continues the flow.

        **Not a bearer credential.** It must never be sent in `Authorization`,
        and it authenticates nothing: no protected route accepts it, and it
        could not be mistaken for an access token even if one tried — the two
        families are hashed under different domains, so neither can validate as
        the other.

        Opaque. It is a lookup key, not a container: clients must not parse it,
        split it, or infer anything from its shape. It expires with the
        registration rather than on a clock of its own, does not rotate, and is
        presented on every step — re-presenting it is continuation, not reuse,
        and revokes nothing. It stops working the moment the registration
        produces an account.

        Store it where the platform keeps secrets for as long as the flow runs,
        and discard it afterwards.
      examples:
        - 'rmreg_00000000-0000-7000-8000-000000000000.EXAMPLE-REGISTRATION-CREDENTIAL-NOT-A-REAL-ONE'
    RegistrationChannel:
      type: string
      description: |
        Which possession a step is about. A registration binds one destination
        per channel, once, and proves each independently — in either order.
      enum: [email, sms]
...
    RegistrationPasscodeRequest:
      type: object
      additionalProperties: false
      required: [registration_credential, channel, destination]
      properties:
        registration_credential:
          $ref: '#/components/schemas/RegistrationCredential'
        channel:
          $ref: '#/components/schemas/RegistrationChannel'
        destination:
          type: string
          minLength: 3
          maxLength: 254
          description: |
            Where the passcode goes — an email address when `channel` is
            `email`, a phone number when it is `sms`.

            Given as the member would write it, and normalized to one canonical
            identity server-side: `0532 123 45 67`, `+90 532 123 45 67` and
            `905321234567` are the same person, and an address is lowercased in
            full. Nothing else is folded — a dot and a `+tag` are part of an
            address, and guessing otherwise merges two people into one identity
            with no way back.

            **This is the only step that accepts one.** The registration binds
            it here and reads it back itself when the passcode is verified.
          examples:
            - 'member@example.com'
...
    RegistrationRefusalReason:
      type: string
      description: |
        Why a registration step was refused, as a stable machine string.

        **Narrower than the server's own vocabulary, deliberately.** The domain
        distinguishes an address already registered from a number already
        registered; this does not, because a caller able to tell them apart
        could read back which half of an identifier pair belongs to a member.
        A credential that has ended is not in this list at all — it is a `401`
        with no reason, so completion history cannot be probed.

        Renaming one is a breaking contract change.
      enum:
        - not_fully_proven
        - account_already_exists
        - channel_already_bound
...
    TripRefusalReason:
      type: string
      enum:
        - recurring_route_unsupported
        - departure_not_reached
        - service_date_passed
        - route_unavailable
        - trip_not_started
        - already_completed
        - already_aborted
      description: |
        Why a trip lifecycle command was refused, as a stable machine string.

        **This is contract**: renaming one is a breaking change. Clients map each
        to their own approved copy, because `message` is developer-facing English
        that no client displays. Every one of these arrives with `409` and
        `conflict` — a trip refusal is always the resource not being in the state
        the command needed, never a malformed request. `departure_not_reached` in
        particular is authoritative mutable state and not a validation failure:
        the same request succeeds later, unchanged.

        `route_unavailable` and `service_date_passed` are also
        `SeatRequestRefusalReason` members, and that is deliberate — the same
        meaning deserves the same wire string. The vocabularies stay separate
        schemas so that neither domain's next reason has to be argued in the
        other.

        `recurring_route_unsupported` no longer means recurring journeys cannot
        be made: Phase 16b starts, completes and abandons them by date. It means
        the caller used the route-only form, which names no day, on a plan that
        has many — so there is nothing to guess from.

        `service_date_passed` bounds **Start only**. Completing or abandoning a
        journey that already exists has no calendar bound: a journey that ran
        past midnight is ended the morning after.

        There is no `trip_already_started`. It could only have described a route
        cancellation refused because a journey had begun, and nothing refuses
        that. Cancelling a **plan** says no future journey of it will run; it
        does not rewrite a journey already under way, and the driver on that one
        must still be able to complete or abort it. (A weekday plan is
        `upcoming` for ever, so cancellation and a running journey genuinely do
        overlap — for a one-off route they never could.) Repeating Start on a
        running trip is a success, not a refusal.

        No reason accompanies an abandoned journey either. `aborted` records that
        the driver said it did not happen; why is a taxonomy nobody has designed.
...
    SeatRequestStatus:
      type: string
      enum: [pending, accepted, declined, withdrawn]
      description: |
        Where one asking stands. Only `pending` can still change.

        There is no `expired`: a request on a journey that has departed is a
        comparison against the clock, not stored state, and persisting it would
        need a sweep to keep it true. There is no `cancelled_by_route` either —
        when a driver withdraws a journey the request keeps saying what the
        member asked and what answer they got, and the route's own `status` says
        the rest.
    SeatRequestRefusalReason:
      type: string
      enum:
        - profile_required
        - own_route
        - id_already_used
        - already_requested
        - route_unavailable
        - service_date_passed
        - route_full
        - already_accepted
        - already_decided
        - withdrawn
      description: |
        Why a seat-request command was refused, as a stable machine string.

        **This is contract**: renaming one is a breaking change. Clients map each
        to their own approved copy, because `message` is developer-facing English
        that no client displays.

        `profile_required` is the only one that arrives with `422` and
        `validation_failed` — the journey is a valid target and the body is well
        formed; what is missing is the caller's own prerequisite. Every other
        reason arrives with `409` and `conflict`.

        There is deliberately no reason for a cancelled or departed journey on
        the create path: those are invisible to a passenger and answer `404`, so
        a distinguishable refusal would confirm which ids exist.
...
    ReviewRefusalReason:
      type: string
      enum:
        - seat_request_not_accepted
        - trip_not_completed
        - review_window_closed
        - already_reviewed
        - id_already_used
      description: |
        Why a review was refused, as a stable machine string.

        **This is contract**: renaming one is a breaking change. All five arrive
        with `409` and `conflict`, so the status distinguishes none of them —
        which is exactly why the reason exists.

        `seat_request_not_accepted` is not reducible to either neighbour. A
        caller is a legitimate party to their own `pending`, `declined` or
        `withdrawn` request, so a 404 would be wrong; and it is independent of
        the journey, since a request can be declined on a route whose trip
        completed, where `trip_not_completed` would be false.

        `id_already_used` is spelled as `SeatRequestRefusalReason` spells it,
        because it means the same thing on the same kind of create path.

        There is deliberately no `self_review` (a member cannot hold an accepted
        request on their own route), no `not_a_participant` (that is the
        non-disclosing 404) and no `profile_required` (a reviewer necessarily
        has a profile). A reason nothing can produce is worse than none.
...